BUG: Fixed most build errors for RHEL6 ENH: Added -r, -c, and -d flags to RHEL6.sh script

+++ scripts/RHEL6.sh
@@ -1,25 +1,35 @@
 #!/bin/bash
-SCRIPTNAME=$(basename ${0} .sh)
+SCRIPTNAME=$(basename ${0})
+DEBUG=false
+BUILD_CMAKE=true
+BUILD_PYTHON=false
+BUILD_QT=true
+BUILD_VIRTUALENV=true
+BUILD_NAMIC=true
 
-BUILD_CMAKE=$(true)
-BUILD_PYTHON=$(false)
-BUILD_QT=$(true)
-BUILD_VENV=$(true)
-BUILD_NAMIC=$(true)
-
-VERBOSE=$(false)
-RUN_TESTS=$(true)
+REFRESH=false
+VERBOSE=false
+RUN_TESTS=false
+CONTINUE=false
 
 function usage ()
 {
-	echo "Usage: ${SCRIPTNAME} [-h] [-v] [-t] " >&2
+	echo "Usage: ${SCRIPTNAME} [-h] [-v] [-t] [-c] [-r] [-d DIR]" >&2
+    echo "" >&2
+    echo "  -h       print this message" >&2
+    echo "  -v       verbose" >&2
+    echo "  -t       testing on" >&2
+    echo "  -c       continue previous build" >&2
+    echo "  -r       refresh/redownload source files" >&2
+    echo "  -d DIR   directory to create/continue (default:'YYYYMMDD')" >&2
     debug
+    echo "Number of inputs is ${#}" >&2
 	[[ ${#} -eq 1 ]] && exit ${1} || exit ${EXIT_FAILURE}
 }
 
 function debug ()
 {
-    if [[ "$DEBUG" == "ON" ]]; then
+    if [[ "${DEBUG}" == true ]]; then
         set | sort > set.sorted
         printenv | sort > printenv.sorted
         printf "\n--------------------DEBUG--------------------\n"
@@ -30,16 +40,16 @@ function debug ()
 
 download_source_code()
 {
-    if $BUILD_CMAKE; then
+    if [[ "${BUILD_CMAKE}" == true ]]; then
         download_cmake
     fi
-    if $BUILD_QT; then
+    if [[ "${BUILD_QT}" == true ]]; then
         download_qt
     fi
-    if $BUILD_VIRTUALENV; then
+    if [[ "${BUILD_VIRTUALENV}" == true ]]; then
         download_virtualenv
     fi
-    if $BUILD_NAMIC; then
+    if [[ "${BUILD_NAMIC}" == true ]]; then
         download_namic
     fi
     return 0
@@ -51,14 +61,27 @@ download_source_code()
 
 build_cmake()
 {
+    if [[ "${CONTINUE}" == true && -f ${MYBUILD}/cmake-bin/bin/cmake ]]; then
+        return 0
+    fi
     echo "Building CMake (version ${CMAKE_VERSION})..."
     cd ${MYBUILD}
     tar -xzf ${MYSOURCE}/cmake-${CMAKE_VERSION}.tar.gz
     cd cmake-${CMAKE_VERSION}
     ./bootstrap --prefix=${MYBUILD}/cmake-bin 1>${MYBUILD}/.log/cmake.stdout 2> ${MYBUILD}/.log/cmake.stderr
-    gmake && gmake install
-    export PATH=${MYBUILD}/cmake-bin/bin:${PATH}
-    echo "Success! Added CMake bin to PATH."
+    wait
+    if make ; then
+        if make install ; then
+            export PATH=${MYBUILD}/cmake-bin/bin:${PATH}
+            echo "Success! Added CMake bin to PATH."
+        else
+            echo "Failed to install CMake! ERROR CODE: $? Halting."
+            return ${ERR_CMAKE_INSTALL}
+        fi
+    else
+        echo "Failed to build CMake! ERROR CODE: $? Halting."
+        return ${ERR_CMAKE_BUILD}
+    fi
     return 0
 }
 
@@ -83,17 +106,42 @@ build_python()
 
 build_qt()
 {
+    if [[ "${CONTINUE}" == true && -f ${MYBUILD}/qt-bin/bin/qmake ]]; then
+        return 0
+    fi
     if ! which Xorg; then  # Verify you have X11 available
         echo "X11 not available on system.  Please try again with X11"
         return -2
     fi
     #NEON: qlogin -pe smp 16 -q HJ
     cd /tmp  # Builds faster on local drive
+    mkdir ${MYBUILD}/qt-bin
     tar -xzf ${MYSOURCE}/qt-everywhere-opensource-src-${QT_VERSION}.tar.gz
     cd qt-everywhere-opensource-src-${QT_VERSION}
-    ./configure --prefix=${MYBUILD}/qt-bin -opensource -qt-zlib -qt-libtiff -qt-libpng -qt-libmng -qt-libjpeg -arch x86_64
-    gmake && gmake install
-    export PATH=${MYBUILD}/qt-bin/bin:${PATH}
+    ./configure --prefix=${MYBUILD}/qt-bin \
+                -opensource \
+                -qt-zlib \
+                -qt-libtiff \
+                -qt-libpng \
+                -qt-libmng \
+                -qt-libjpeg \
+                -arch x86_64 \
+                <<QTLicenseAgree
+yes
+QTLicenseAgree
+    wait
+    if gmake ; then
+        if gmake install ; then
+            export PATH=${MYBUILD}/qt-bin/bin:${PATH}
+            return 0
+        else
+            echo "Failed to install Qt. ERROR CODE: $? Halting."
+            return ${ERR_QT_INSTALL}
+        fi
+    else
+        echo "Failed to build Qt. ERROR CODE: $? Halting."
+        return ${ERR_QT_BUILD}
+    fi
     return 0
 }
 
@@ -105,16 +153,29 @@ build_virtualenv()
         MYENV=$1
     fi
     cd ${MYBUILD}
-    tar -xzf ${MYSOURCE}/virtualenv-${VENV_VERSION}.tar.gz
-    python ${MYBUILD}/virtualenv-${VENV_VERSION}/virtualenv.py --always-copy ${MYBUILD}/${MYENV}
-    ln -s ${MYBUILD}/${MYENV}/bin/activate ${MYBUILD}/activate_${MYENV}
-    # Medusa: no BLAS library!
+    if tar -xzf ${MYSOURCE}/virtualenv-${VENV_VERSION}.tar.gz ; then
+        python ${MYBUILD}/virtualenv-${VENV_VERSION}/virtualenv.py --always-copy ${MYBUILD}/${MYENV}
+        wait
+    else
+        echo "Could not uncompress virtualenv package!  Halting."
+        return ${ERR_MISSING_DIR}
+    fi
+    if [[ -f ${MYBUILD}/${MYENV}/bin/activate ]]; then
+        ln -s ${MYBUILD}/${MYENV}/bin/activate ${MYBUILD}/activate_${MYENV}
+    else
+        return ${ERR_MISSING_FILE}
+    fi
+    # On Medusa: no BLAS library!
     # build_ATLAS()
     source ${MYBUILD}/activate_${MYENV}
     pip install nose python-dateutil six wsgiref
+    wait
     pip install numpy scipy matplotlib tornado pyzmq ipython
+    wait
     pip install networkx nibabel traits
+    wait
     pip install pydicom
+    wait
     return 0
 }
 
@@ -134,15 +195,26 @@ build_NEP()
 {
    mkdir ${MYBUILD}/NAMIC-build
    cd ${MYBUILD}/NAMIC-build
+   source ${MYBUILD}/activate_$1
    CC=/usr/bin/gcc \
-     CXX=/usr/bin/g++ \
-     ccmake ${MYSOURCE}/NAMICExternalProjects/ \
-     -DQT_QMAKE_EXECUTABLE:FILEPATH=${MYBUILD}/qt-bin/bin/qmake \
-     -DPYTHON_EXECUTABLE:FILEPATH=$(which python) \
-     -CMAKE_INSTALL_PREFIX:FILEPATH=${MYBUILD} \
-     -DSITK_INT64_PIXELIDS:BOOL=OFF  # SimpleITK flag
-   make -j 12 -k
-   make install
+   CXX=/usr/bin/g++ \
+   cmake ${MYSOURCE}/NAMICExternalProjects/ \
+    -DQT_QMAKE_EXECUTABLE:FILEPATH=${MYBUILD}/qt-bin/bin/qmake \
+    -DPYTHON_EXECUTABLE:FILEPATH=$(which python) \
+    -CMAKE_INSTALL_PREFIX:FILEPATH=${MYBUILD} \
+    -DSITK_INT64_PIXELIDS:BOOL=OFF  # SimpleITK flag
+   wait
+   if make -j 16 -k ; then
+       if make install ; then
+           continue
+       else
+           echo "Failed to install NAMIC. ERROR CODE: $? Halting."
+           return ${ERR_NAMIC_INSTALL}
+       fi
+   else
+       echo "Failed to build NAMIC. ERROR CODE: $? Halting."
+       return ${ERR_NAMIC_BUILD}
+   fi
    echo "\
 # --- SINAPSE-specific modifications ---
 PYTHONPATH=${MYBUILD}/NAMIC-build/SimpleITK-build/Wrapping/:${MYBUILD}/NAMIC-build/NIPYPE \
@@ -181,14 +253,21 @@ test_sitk()
 }
 
 # MAIN
-while getopts ':o:vht' OPTION ; do
-    debug
+source codes.sh
+while getopts ':o:vhd:tcr' OPTION ; do
+    # debug
 	case ${OPTION} in
-		v)	VERBOSE=$(true)
+		v)	VERBOSE=true
 			;;
-		h)	usage ${EXIT_SUCCESS}
+		h)	usage 0
 			;;
-        t)  RUN_TESTS=$(true)
+        t)  RUN_TESTS=true
+            ;;
+        c)  CONTINUE=true
+            ;;
+        d)  export DATETAG=${OPTARG}  # TODO: verify correct date formatting
+            ;;
+        r)  REFRESH=true
             ;;
 		\?)	echo "unknown option \"-${OPTARG}\"." >&2
 			usage ${EXIT_ERROR}
@@ -201,37 +280,67 @@ while getopts ':o:vht' OPTION ; do
 			;;
 	esac
 done
+
 source creator.sh
-source $PWD/scripts/download.sh
+source scripts/download.sh
 
-if [[ -d ${MYSOURCE} ]]; then
-    rm -rf ${MYSOURCE}  # TODO: Add option flag
-fi
-mkdir -p ${MYSOURCE}
-if [[ ! -d ${MYSOURCE} ]]; then
-    exit ${ERR_MISSING_DIR}
+if [[ "$CONTINUE" == true ]]; then
+    if [[ ! -d ${BASEDIR}/${DATETAG} ]]; then
+        echo "Directory ${DATETAG} not found in ${BASEDIR}!" >&2
+        usage ${ERR_MISSING_DIR}
+    fi
 fi
 
-if [[ -d ${MYBUILD} ]]; then
-    rm -rf ${MYBUILD}  # TODO: Add option flag
+if [[ "${CONTINUE}" == "false" ]]; then
+    if [[ -d ${MYSOURCE} ]]; then
+        #HACK
+        exit 100
+        mv ${MYSOURCE} ${BASEDIR}/.xxx
+        rm -rf ${BASEDIR}/xxx  # TODO: Add option flag
+    fi
+    mkdir -p ${MYSOURCE}
+    if [[ -d ${MYBUILD} ]]; then
+        #HACK
+        exit 100
+        mv ${MYBUILD} ${BASEDIR}/.yyy
+        rm -rf ${BASEDIR}/.yyy  # TODO: Add option flag
+    fi
+    mkdir -p ${MYBUILD}/.log
 fi
-mkdir -p ${MYBUILD}/.log
-if [[ ! -d ${MYBUILD}/.log ]]; then
+
+if [[ ! -d ${MYSOURCE} || ! -d ${MYBUILD}/.log ]]; then
+    echo "Cannot find $MYSOURCE or $MYBUILD! Halting."
     exit ${ERR_MISSING_DIR}
 fi
 
 if ! module load python/2.7; then
     exit ${ERR_MODULE_LOAD}
 fi
-if download_source_code ; then
-    build_cmake;                     CMAKE_STATUS=$?
-    build_qt;                        QT_STATUS=$?
-    build_python;                    PYTHON_STATUS=$?
+
+if [[ "${CONTINUE}" == false || "${REFRESH}" == true ]]; then
+    download_source_code  # TODO: download source code if not there OR 'refresh' option
+fi
+
+if [[ true ]]; then
+    echo "Building CMake..."
+    build_cmake
+    CMAKE_STATUS=$?
+    echo "Building Qt..."
+    build_qt
+    QT_STATUS=$?
+    echo "Building Python..."
+    build_python
+    PYTHON_STATUS=$?
+    echo "Creating virtualenv..."
     MYENV=python_HD
-    build_virtualenv ${MYENV};    VENV_STATUS=$?
-    if [[ expr ${VENV_STATUS} + ${CMAKE_STATUS} + ${QT_STATUS} + ${PYTHON_STATUS} ]]; then
-        source ${MYBUILD}/activate_${MYENV}
-        build_NEP;                   NAMIC_STATUS=$?
+    build_virtualenv ${MYENV}
+    VENV_STATUS=$?
+    wait
+    if [[ $(expr ${VENV_STATUS} + ${CMAKE_STATUS} + ${QT_STATUS} + ${PYTHON_STATUS}) -eq 0 ]]; then
+        echo "Building NAMICExternalProjects..."
+        build_NEP ${MYENV}
+        NAMIC_STATUS=$?
+        wait
         if ${NAMIC_STATUS}; then
             deactivate
             if ${RUN_TESTS}; then
